requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.1
django-admin-sortable2==2.1.3
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-tinymce==3.5.0
djangorestframework==3.14.0
Faker==15.1.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1